refactor: 디테일 모달 컴포넌트 분리
