+++ README.md.md
@@ -57,7 +57,7 @@
 
 피드백 바탕 리팩토링 사항
 
-- [ ] 디테일 모달 컴포넌트 분리
+- [x] 디테일 모달 컴포넌트 분리
 - [x] starRating 5 상수 처리
 - [x] 별점 소수점 처리
 - [ ] 모달이나 바텀 시트가 열렸을 때 백드롭 영역 스크롤 막기
